refactor(client-cli): extract cdb download v1 execution code to a `v1` module
